Cleaning up sample
