+++ sample/build.gradle
@@ -23,7 +23,8 @@ android {
             annotationProcessorOptions {
                 arguments = [
                         stagGeneratedPackageName: 'com.vimeo.sample.stag.generated',
-                        stagDebug               : 'true'
+                        stagDebug               : 'true',
+                        stagAssumeHungarianNotation: 'true'
                 ]
             }
         }
