+++ sample/src/main/java/com/vimeo/sample/model/IdenticalFieldTypes.java
@@ -24,6 +24,8 @@
 package com.vimeo.sample.model;
 
 import com.google.gson.annotations.SerializedName;
+import com.vimeo.sample.model.root.Stats;
+import com.vimeo.sample.model.root.User;
 import com.vimeo.stag.UseStag;
 
 import java.util.ArrayList;
