+++ sample/src/main/java/com/vimeo/sample/model/TestDeserializer.java
@@ -4,6 +4,7 @@ import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParseException;
+import com.vimeo.sample.model.root.User;
 
 import java.lang.reflect.Type;
