+++ sample/src/main/java/com/vimeo/sample/model/TestSerializer.java
@@ -4,6 +4,7 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSerializationContext;
 import com.google.gson.JsonSerializer;
+import com.vimeo.sample.model.root.User;
 
 import java.lang.reflect.Type;
