+++ sample/src/main/java/com/vimeo/sample/model/TestSerializerDeserializer.java
@@ -7,6 +7,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 import com.google.gson.JsonSerializationContext;
 import com.google.gson.JsonSerializer;
+import com.vimeo.sample.model.root.User;
 
 import java.lang.reflect.Type;
