--- sample/src/main/java/com/vimeo/sample/model1/Video.java
@@ -1,34 +1,0 @@
-package com.vimeo.sample.model1;
-
-import android.support.annotation.NonNull;
-
-import com.google.gson.annotations.SerializedName;
-import com.vimeo.sample.model.User;
-import com.vimeo.stag.UseStag;
-
-/**
- * This class has been added to reproduce the scenario where two classes with same
- * names exists in different packages. This will make sure that the getter type
- * adapter method for these classes are generated with the correct name (appending
- * package name) so that two getters with same name is not created.
- * <p>
- * This class has been referenced as a member variable in {@link com.vimeo.sample.model.Video} class
- */
-@UseStag
-public class Video {
-
-    @NonNull
-    @SerializedName("user")
-    public User mUser;
-
-    @NonNull
-    @SerializedName("link")
-    public String mLink;
-
-    @NonNull
-    @SerializedName("mData")
-    public Data mData;
-
-    @SerializedName("mParameterizedData")
-    public ParameterizedData<String> mParameterizedData;
-}
