--- sample/src/test/java/com/vimeo/sample/model/VideoListTest.java
@@ -1,17 +1,0 @@
-package com.vimeo.sample.model;
-
-import com.vimeo.sample.Utils;
-
-import org.junit.Test;
-
-/**
- * Created by restainoa on 2/2/17.
- */
-public class VideoListTest {
-
-    @Test
-    public void typeAdapterWasGenerated() throws Exception {
-        Utils.verifyTypeAdapterGeneration(VideoList.class);
-    }
-
-}
