+++ sample/src/test/java/com/vimeo/sample/model/root/PagingTest.java
@@ -1,17 +1,18 @@
-package com.vimeo.sample.model;
+package com.vimeo.sample.model.root;
 
 import com.vimeo.sample.Utils;
+import com.vimeo.sample.model.root.Paging;
 
 import org.junit.Test;
 
 /**
  * Created by restainoa on 2/2/17.
  */
-public class UserTest {
+public class PagingTest {
 
     @Test
     public void typeAdapterWasGenerated() throws Exception {
-        Utils.verifyTypeAdapterGeneration(User.class);
+        Utils.verifyTypeAdapterGeneration(Paging.class);
     }
 
 }
