+++ sample/src/test/java/com/vimeo/sample/model/root/StatsTest.java
@@ -1,17 +1,18 @@
-package com.vimeo.sample.model;
+package com.vimeo.sample.model.root;
 
 import com.vimeo.sample.Utils;
+import com.vimeo.sample.model.root.Stats;
 
 import org.junit.Test;
 
 /**
  * Created by restainoa on 2/2/17.
  */
-public class VideoTest {
+public class StatsTest {
 
     @Test
     public void typeAdapterWasGenerated() throws Exception {
-        Utils.verifyTypeAdapterGeneration(Video.class);
+        Utils.verifyTypeAdapterGeneration(Stats.class);
     }
 
 }
