+++ sample/src/test/java/com/vimeo/sample/model/root/UserTest.java
@@ -1,17 +1,18 @@
-package com.vimeo.sample.model;
+package com.vimeo.sample.model.root;
 
 import com.vimeo.sample.Utils;
+import com.vimeo.sample.model.root.User;
 
 import org.junit.Test;
 
 /**
  * Created by restainoa on 2/2/17.
  */
-public class StatsTest {
+public class UserTest {
 
     @Test
     public void typeAdapterWasGenerated() throws Exception {
-        Utils.verifyTypeAdapterGeneration(Stats.class);
+        Utils.verifyTypeAdapterGeneration(User.class);
     }
 
 }
