+++ sample/src/test/java/com/vimeo/sample/model/root/VideoTest.java
@@ -1,17 +1,18 @@
-package com.vimeo.sample.model;
+package com.vimeo.sample.model.root;
 
 import com.vimeo.sample.Utils;
+import com.vimeo.sample.model.root.Video;
 
 import org.junit.Test;
 
 /**
  * Created by restainoa on 2/2/17.
  */
-public class PagingTest {
+public class VideoTest {
 
     @Test
     public void typeAdapterWasGenerated() throws Exception {
-        Utils.verifyTypeAdapterGeneration(Paging.class);
+        Utils.verifyTypeAdapterGeneration(Video.class);
     }
 
 }
