+++ sample/src/test/java/com/vimeo/sample/model1/DuplicateNameTest.java
@@ -1,0 +1,16 @@
+package com.vimeo.sample.model1;
+
+import com.vimeo.sample.Utils;
+
+import org.junit.Test;
+
+/**
+ * Unit tests for {@link DuplicateName}.
+ */
+public class DuplicateNameTest {
+
+    @Test
+    public void verifyTypeAdapterWasGenerated() throws Exception {
+        Utils.verifyTypeAdapterGeneration(DuplicateName.class);
+    }
+}
